add perf log for hexagon-npu backend
